+++ pycbc/conversions.py
